Refactor test
* Add test_helper and Rakefile
* Rename test in accordance to unit-test typical rule.
