+++ ruby/Gemfile
@@ -10,3 +10,4 @@ gem 'red-datasets'
 gem 'test-unit'
 gem 'irb'
 gem 'pry-byebug'
+gem 'rake'
